fix #29
